Downgrade to mongodb node driver 4.0.1
Version 4.8.x (the first minor version that works with Mongo 6) is
causing errors we don't understand (failing to find the Mongo
replicaset, although we can access it easily using the mongo shell).

Rather than delve into that, we're going to try version 4.0.x first,
which is the first version to work with Mongo 5. It will mean we'll have
to figure out the other problem at some point before the election when
we upgrade to Mongo 6, but we don't have to deal with it now.