Upgrade mongo driver from 3.6.1 to 4.3.1
We need to upgrade to 4.x.x because 3.x.x isn't compatible with
Mongo 7.

We're choosing 4.3.1 because it's the version the parse-server
community project upgraded to when they were preparing for Mongo 5.
We are about to cherry-pick a few changes from that commit in the
next couple commits, so we're hoping that if the versions match,
it will work (4.8.1 was too much, apparently).

Here's the commit where they made that upgrade:
https://github.com/Hustle/parse-server/commit/f88aa2a62a533e5344d1c13dd38c5a0b283a480a

Note: We're upgrading to Mongo 7 but 4.3 (a pretty old version) will still
work for us, even though not all the Mongo 7 features can be used.
(See https://www.mongodb.com/docs/drivers/node/v4.3/compatibility/). We'll
try to upgrade further once we're not under the Mongo 7 forced upgrade deadline.